В ключена локализация интерфейса. Рефакторинг
