--- data/questions.csv
@@ -1,6 +1,0 @@
-id,text,answers,trueAns
-1,Выбирете допустимые объявления java, int i=0;: int i=125;: int i="asd";: Int i==125; : float r=125f;,2:1:5
-2,Как выглядит оператор присваивания в java, i eq 125 ;: abc<=>125;: int i=="asd";: i=125;: f=125f;,4:5
-3,Как выглядит оператор сравнения в java, i eq 125 ;: abc<=>125;: i==12;:i=125; :Int=125.00;,3
-4,Как выглядит оператор НЕ РАВНО в java, i <> 125 ;: abc >< 125;: i==12;: i~=125; : jjj != kkk;,5
-5,Что делает операция ++ в java, увеличивает объем памяти : ничего ни делает : возращает строку: увеличивает счетчик :  уменьшает счетчик,4
