--- data/students.csv
@@ -1,4 +1,0 @@
-id,name,password
-1,Ivanov,1
-2,petrov,22
-3,sidorov,333
