--- src/main/java/cvr/otus/GLB_CONST.java
@@ -1,13 +1,0 @@
-package cvr.otus;
-
-/**
- * Глобальные кконстанты
- */
-public final class GLB_CONST {
-    public final static String PRG_CAPT = "program.caption";
-    public final static String STUDENT_NAME = "student.name";
-    public final static String STUDENT_PASSWORD = "student.password";
-    public final static String TRY_AGAIN= "try.again";
-    public final static String ATTEMPTS_EXHAUSTED= "attempts.exhausted";
-    //Attempts exhausted
-}
