+++ src/main/java/cvr/otus/Main.java
@@ -2,9 +2,8 @@ package cvr.otus;
 
 import cvr.otus.domain.Student;
 import cvr.otus.service.ExamRunner;
+import cvr.otus.service.ExamRunnerImpl;
 import cvr.otus.service.LoginService;
-import cvr.otus.service.LoginServiceImpl;
-import cvr.otus.service.StudentService;
 import cvr.otus.utils.PrintService;
 import cvr.otus.utils.PrintServiceImpl;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
@@ -13,8 +12,6 @@ import org.springframework.context.annotation.PropertySource;
 import org.springframework.context.support.AbstractApplicationContext;
 
 import java.util.Scanner;
-
-//import static cvr.otus.GLB_CONST.PRG_CAPT;
 
 @ComponentScan
 @PropertySource("classpath:file-path.properties")
@@ -27,21 +24,16 @@ public class Main {
         PrintService ps = context.getBean(PrintServiceImpl.class);
         ps.sayln("program.caption");
 
-
-        StudentService studentService = context.getBean(StudentService.class);
         LoginService loginService = context.getBean(LoginService.class);
-//        LoginService loginService = new LoginServiceImpl(studentService, ps);
-
         loginService.setScanner(new Scanner(System.in));
         Student student = loginService.login();
 
         if (student != null) {
-            ExamRunner examRunner = context.getBean(ExamRunner.class);
+            ExamRunner examRunner = context.getBean(ExamRunnerImpl.class);
             examRunner.setStudent(student);
             examRunner.setScanner(new Scanner(System.in));
             examRunner.run();
         }
-
     }
