+++ src/main/java/cvr/otus/data/DataLoader.java
@@ -1,6 +1,5 @@
 package cvr.otus.data;
 
-import cvr.otus.data.Data;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Component;
