+++ src/main/resources/file-path.properties
@@ -3,6 +3,6 @@ students=./data/students.csv
 
 # Язык
 # Русский
-language=ru,RU 
+language=ru,RU
 #Албанский
 #language=al,AL
