+++ src/main/resources/i18n/bundle_al.properties
@@ -1,3 +1,4 @@
+# Бандл на албанском языке
 str={0}
 program.caption=Programi i Testimit të Studentëve
 student.name=Emri yt    :
