+++ src/test/java/cvr/otus/service/QuestionServiceImplTest.java
@@ -4,7 +4,6 @@ import cvr.otus.dao.QuestionDao;
 import cvr.otus.dao.QuestionDaoImplTest;
 import cvr.otus.domain.Question;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.context.annotation.PropertySource;
 
@@ -13,7 +12,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
